Cleaned up member filters
fixes https://github.com/TryGhost/Team/issues/2134

Moved all filters to separate files to make the filter component a lot more readable and easier to maintain.
