+++ ghost/admin/app/utils/member-filters/created-at.js
@@ -1,0 +1,8 @@
+import {DATE_RELATION_OPTIONS} from './relation-options';
+
+export const CREATED_AT_FILTER = {
+    label: 'Created', 
+    name: 'created_at', 
+    valueType: 'date', 
+    relationOptions: DATE_RELATION_OPTIONS
+};
