+++ ghost/admin/app/utils/member-filters/email-count.js
@@ -1,0 +1,9 @@
+import {NUMBER_RELATION_OPTIONS} from './relation-options';
+
+export const EMAIL_COUNT_FILTER = {
+    label: 'Emails sent (all time)', 
+    name: 'email_count', 
+    columnLabel: 'Email count', 
+    valueType: 'number', 
+    relationOptions: NUMBER_RELATION_OPTIONS
+};
