+++ ghost/admin/app/utils/member-filters/email-open-rate.js
@@ -1,0 +1,8 @@
+import {NUMBER_RELATION_OPTIONS} from './relation-options';
+
+export const EMAIL_OPEN_RATE_FILTER = {
+    label: 'Open rate (all time)', 
+    name: 'email_open_rate', 
+    valueType: 'number', 
+    relationOptions: NUMBER_RELATION_OPTIONS
+};
