+++ ghost/admin/app/utils/member-filters/email-opened-count.js
@@ -1,0 +1,9 @@
+import {NUMBER_RELATION_OPTIONS} from './relation-options';
+
+export const EMAIL_OPENED_COUNT_FILTER = {
+    label: 'Emails opened (all time)', 
+    name: 'email_opened_count', 
+    columnLabel: 'Email opened count', 
+    valueType: 'number', 
+    relationOptions: NUMBER_RELATION_OPTIONS
+};
