+++ ghost/admin/app/utils/member-filters/email.js
@@ -1,0 +1,8 @@
+import {CONTAINS_RELATION_OPTIONS} from './relation-options';
+
+export const EMAIL_FILTER = {
+    label: 'Email', 
+    name: 'email',
+    valueType: 'string', 
+    relationOptions: CONTAINS_RELATION_OPTIONS
+};
