+++ ghost/admin/app/utils/member-filters/label.js
@@ -1,0 +1,9 @@
+import {MATCH_RELATION_OPTIONS} from './relation-options';
+
+export const LABEL_FILTER = {
+    label: 'Label', 
+    name: 'label', 
+    valueType: 'array', 
+    columnLabel: 'Label', 
+    relationOptions: MATCH_RELATION_OPTIONS
+};
