+++ ghost/admin/app/utils/member-filters/last-seen.js
@@ -1,0 +1,9 @@
+import {DATE_RELATION_OPTIONS} from './relation-options';
+
+export const LAST_SEEN_FILTER = {
+    label: 'Last seen', 
+    name: 'last_seen_at', 
+    valueType: 'date', 
+    columnLabel: 'Last seen at', 
+    relationOptions: DATE_RELATION_OPTIONS
+};
