+++ ghost/admin/app/utils/member-filters/name.js
@@ -1,0 +1,8 @@
+import {CONTAINS_RELATION_OPTIONS} from './relation-options';
+
+export const NAME_FILTER = {
+    label: 'Name', 
+    name: 'name', 
+    valueType: 'string', 
+    relationOptions: CONTAINS_RELATION_OPTIONS
+};
