+++ ghost/admin/app/utils/member-filters/next-billing-date.js
@@ -1,0 +1,9 @@
+import {DATE_RELATION_OPTIONS} from './relation-options';
+
+export const NEXT_BILLING_DATE_FILTER = {
+    label: 'Next billing date', 
+    name: 'subscriptions.current_period_end', 
+    valueType: 'date', 
+    columnLabel: 'Next billing date', 
+    relationOptions: DATE_RELATION_OPTIONS
+};
