+++ ghost/admin/app/utils/member-filters/relation-options/contains.js
@@ -1,0 +1,7 @@
+export const CONTAINS_RELATION_OPTIONS = [
+    {label: 'is', name: 'is'},
+    {label: 'contains', name: 'contains'},
+    {label: 'does not contain', name: 'does-not-contain'},
+    {label: 'starts with', name: 'starts-with'},
+    {label: 'ends with', name: 'ends-with'}
+];
