+++ ghost/admin/app/utils/member-filters/relation-options/date.js
@@ -1,0 +1,6 @@
+export const DATE_RELATION_OPTIONS = [
+    {label: 'before', name: 'is-less'},
+    {label: 'on or before', name: 'is-or-less'},
+    {label: 'after', name: 'is-greater'},
+    {label: 'on or after', name: 'is-or-greater'}
+];
