+++ ghost/admin/app/utils/member-filters/relation-options/match.js
@@ -1,0 +1,4 @@
+export const MATCH_RELATION_OPTIONS = [
+    {label: 'is', name: 'is'},
+    {label: 'is not', name: 'is-not'}
+];
