+++ ghost/admin/app/utils/member-filters/relation-options/number.js
@@ -1,0 +1,5 @@
+export const NUMBER_RELATION_OPTIONS = [
+    {label: 'is', name: 'is'},
+    {label: 'is greater than', name: 'is-greater'},
+    {label: 'is less than', name: 'is-less'}
+];
