+++ ghost/admin/app/utils/member-filters/subscription-start-date.js
@@ -1,0 +1,9 @@
+import {DATE_RELATION_OPTIONS} from './relation-options';
+
+export const SUBSCRIPTION_START_DATE_FILTER = {
+    label: 'Paid start date', 
+    name: 'subscriptions.start_date', 
+    valueType: 'date', 
+    columnLabel: 'Paid start date', 
+    relationOptions: DATE_RELATION_OPTIONS
+};
