+++ ghost/admin/app/utils/member-filters/tier.js
@@ -1,0 +1,9 @@
+import {MATCH_RELATION_OPTIONS} from './relation-options';
+
+export const TIER_FILTER = {
+    label: 'Membership tier', 
+    name: 'tier', 
+    valueType: 'array', 
+    columnLabel: 'Membership tier', 
+    relationOptions: MATCH_RELATION_OPTIONS
+};
